% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsaArbones-package.R
\docType{package}
\name{gsaArbones-package}
\alias{gsaArbones}
\alias{gsaArbones-package}
\title{gsaArbones: Implements the Fisher Combined Method}
\description{
Consider a set of k independent tests, each of these to test a certain null hypothesis H0|i, i = {1, 2, . . . , k}. For each test, a significance level Pi, i.e., a p-value, is obtained. All these p-values can be combined into a joint test whether there is a global effect, i.e., if a global null hypothesis H0 can be rejected. Gene-set analysis (GSA) algorithms allow to assess the overall evidence of association of a set of genes with a phenotype by combining marginal pvalues of a set of genetic variants (SNPs).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arbones/gsaArbones}
  \item Report bugs at \url{https://github.com/arbones/gsaArbones/issues}
}

}
\author{
\strong{Maintainer}: Jose M. Arbones-Mainar \email{jmarbones.iacs@aragon.es} (\href{https://orcid.org/0000-0002-8982-3737}{ORCID})

}
\keyword{internal}
