% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAfisher.R
\name{GSAfisher.multiple}
\alias{GSAfisher.multiple}
\alias{GSAfisher-multiple}
\title{Gene-Set Analysis Using Fisher's Method for a list of vectors}
\usage{
GSAfisher.multiple(p_values_list)
}
\arguments{
\item{p_values_list}{A list of p-values}
}
\value{
Returns a vector with the p-value of all Fisher's combined statistic tests.
}
\description{
The default method for GSAfisher, handling a single vector of p-values.

This function performs Fisher's combined probability test on a list of p-values.
It can handle a list of p-values.
}
\examples{
p_values_matrix <- matrix(replicate(500,runif(100, 0, 1)),byrow = TRUE,nrow = 500)
p_values_list <- split(p_values_matrix,row(p_values_matrix))
result_multiple <- GSAfisher.multiple(p_values_list)
plot(result_multiple)
}
