% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAfisher.R
\name{GSAfisher}
\alias{GSAfisher}
\alias{GSAfisher-default}
\title{Gene-Set Analysis Using Fisher's Method}
\usage{
GSAfisher(p_values)
}
\arguments{
\item{p_values}{A vector of p-values
This function performs Fisher's combined probability test on a vector or list of p-values.}
}
\value{
Returns a list containing Fisher's combined statistic and the combined p-value. It also returns a vector of all combined p-values.
}
\description{
The default method for GSAfisher, handling a single vector of p-values.
}
\examples{
p_values_single <- runif(10, 0, 0.25)
result_single <- GSAfisher(p_values_single)
print(result_single)
}
