% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAfisher.R
\name{plot.GSAfisherMultipleResult}
\alias{plot.GSAfisherMultipleResult}
\title{Plot of GSAfisher.multiple Analysis}
\usage{
\method{plot}{GSAfisherMultipleResult}(x, ...)
}
\arguments{
\item{x}{An object of class \code{GSAfisherMultipleResult}.}

\item{...}{Other plot arguments}
}
\value{
An histogram of all combined p-values
}
\description{
Generates a concise summary of the GSAfisher analysis results.
}
