% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAfisher.R
\name{print.GSAfisherSummary}
\alias{print.GSAfisherSummary}
\title{Print GSAfisher Results}
\usage{
\method{print}{GSAfisherSummary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{GSAfisherSummary}.}

\item{...}{other arguments}
}
\value{
Summary.
}
\description{
Prints the GSAfisher analysis results directly to the console.
}
