% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAfisher.R
\name{summary.GSAfisherResult}
\alias{summary.GSAfisherResult}
\title{Summary of GSAfisher Analysis}
\usage{
\method{summary}{GSAfisherResult}(object, ...)
}
\arguments{
\item{object}{An object of class \code{GSAfisherResult}.}

\item{...}{other arguments}
}
\value{
Invisible returns a GSAfisherSummary object with key statistics from the Fisher's combined test.
}
\description{
Generates a concise summary of the GSAfisher analysis results.
}
