% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAfisher.R
\name{print.GSAfisherResult}
\alias{print.GSAfisherResult}
\title{Print GSAfisher Results}
\usage{
\method{print}{GSAfisherResult}(object)
}
\arguments{
\item{object}{An object of class \code{GSAfisherResult}.}
}
\value{
returns the input object.
}
\description{
Prints the GSAfisher analysis results directly to the console.
}
