% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSAfisher.R
\name{print.GSAfisherSummary}
\alias{print.GSAfisherSummary}
\title{Print GSAfisher Results}
\usage{
\method{print}{GSAfisherSummary}(object)
}
\arguments{
\item{object}{An object of class \code{GSAfisherSummary}.}
}
\value{
Summary.
}
\description{
Prints the GSAfisher analysis results directly to the console.
}
